+++ Engine/EffectInstance.cpp
@@ -4515,11 +4515,11 @@ EffectInstance::getAvailableLayers(double time, ViewIdx view, int inputNb, std::
 
 
         EffectInstance::ComponentsNeededMap comps;
-        double passThroughTime;
-        int passThroughView;
-        int passThroughInputNb;
+        double passThroughTime = 0.;
+        int passThroughView = 0;
+        int passThroughInputNb = -1; // prevent infinite recursion, because getComponentsNeededAndProduced_public() may call getAvailableLayers()
         std::bitset<4> processChannels;
-        bool processAll;
+        bool processAll = false;
         effect->getComponentsNeededAndProduced_public(getRenderHash(), time, view, &comps, &passThroughLayers, &processAll, &passThroughTime, &passThroughView, &processChannels, &passThroughInputNb);
 
         // Merge pass-through planes produced + pass-through available planes and make it as the pass-through planes for this node